refactor 'assertEquals'
* merge 'testing/pretty.ts' into 'testing/asserts.ts'

* throw AssertionError in assertEquals

* update misc tests use AssertionError
