+++ testing/test.ts
@@ -9,7 +9,6 @@ import {
 } from "./asserts.ts";
 import "./format_test.ts";
 import "./diff_test.ts";
-import "./pretty_test.ts";
 import "./asserts_test.ts";
 import "./bench_test.ts";
